bug fixes and check connection permissions

+++ packages/nip47/lib/src/domain/entities/multi_pay_invoice_request_invoice.dart
@@ -9,7 +9,7 @@ sealed class MultiPayInvoiceRequestInvoice
   const factory MultiPayInvoiceRequestInvoice({
     String? id,
     required String invoice,
-    BigInt? amountSat,
+    int? amountSat,
   }) = _MultiPayInvoiceRequestInvoice;
   const MultiPayInvoiceRequestInvoice._();
